# backend2/src/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BBWS Sumatera VIII API Reference",
    "version": "1.0.0",
    "description": "Dokumentasi REST API BBWS Sumatera VIII Hono Engine (Auto-generated)"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Server Lokal"
    }
  ],
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
...
    "/balai/bbwssumatera8/api/infografis": {
      "get": {
        "tags": [
          "Infografis"
        ],
        "summary": "GET /balai/bbwssumatera8/api/infografis",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        }
      }
    },
    "/balai/bbwssumatera8/api/infografis/{id}": {
      "get": {
        "tags": [
          "Infografis"
        ],
        "summary": "GET /balai/bbwssumatera8/api/infografis/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        }
      },
      "put": {
        "tags": [
          "Infografis"
        ],
        "summary": "PUT /balai/bbwssumatera8/api/infografis/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Infografis"
        ],
        "summary": "DELETE /balai/bbwssumatera8/api/infografis/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/balai/bbwssumatera8/api/infografis/upload": {
      "post": {
        "tags": [
          "Infografis"
        ],
        "summary": "POST /balai/bbwssumatera8/api/infografis/upload",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
    "/balai/bbwssumatera8/api/galeri": {
      "get": {
        "tags": [
          "Galeri"
        ],
        "summary": "GET /balai/bbwssumatera8/api/galeri",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        }
      }
    },
    "/balai/bbwssumatera8/api/galeri/{id}": {
      "get": {
        "tags": [
          "Galeri"
        ],
        "summary": "GET /balai/bbwssumatera8/api/galeri/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        }
      },
      "put": {
        "tags": [
          "Galeri"
        ],
        "summary": "PUT /balai/bbwssumatera8/api/galeri/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Galeri"
        ],
        "summary": "DELETE /balai/bbwssumatera8/api/galeri/{id}",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
    "/balai/bbwssumatera8/api/galeri/upload": {
      "post": {
        "tags": [
          "Galeri"
        ],
        "summary": "POST /balai/bbwssumatera8/api/galeri/upload",
        "responses": {
          "200": {
            "description": "Respons Berhasil",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "401": {
            "description": "Tidak terautentikasi (Unauthorized)"
          },
          "404": {
            "description": "Resource Tidak Ditemukan"
          },
          "500": {
            "description": "Kesalahan Server Internal"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ]
      }
    },
...
  }
}